Adhust Forn,  VM, Fix Model. Minimal functionality for manual calculations works

+++ TaxHelper/Models/PaymentModel.cs
@@ -5,7 +5,7 @@ namespace TaxHelper.Models
     public class PaymentModel
     {
         public DateTime PaymentDate { get; set; }
-        public int PaymentSum { get; set; }
+        public double PaymentSum { get; set; }
         public Currencies PaymentCurrency { get; set; }
     }
 }
